Conflicts:
	buildtools/Cmake/DefinePackages.cmake
	include/surf/simgrid_dtd.h
	include/xbt/graphxml.h
	src/instr/instr_config.c
	src/instr/instr_private.h
	src/simdag/dax_dtd.c
	src/simdag/dax_dtd.h
	src/surf/simgrid_dtd.c
	src/xbt/graphxml.c

+++ teshsuite/smpi/mpich-test/pt2pt/typelb.c
@@ -26,13 +26,14 @@ main( int argc, char **argv)
 	
 	/* Check that the results are correct */
 #ifdef DEBUG
-	printf("lb=%ld, ub=%ld, extent=%ld\n", lb, ub, extent);
+	printf("lb=%ld, ub=%ld, extent=%ld\n",
+               (long)lb, (long)ub, (long)extent);
 	printf("Should be lb=4, ub=5, extent=1\n");
 #endif
 	if (lb != 4 || ub != 5 || extent != 1) {
-	    printf ("lb = %ld (should be 4), ub = %ld (should be 5) extent = %ld should be 1\n", lb, ub, extent) ;
-	}
-	else {
+            printf("lb = %ld (should be 4), ub = %ld (should be 5) extent = %ld should be 1\n",
+                   (long)lb, (long)ub, (long)extent) ;
+	} else {
 	    printf( " No Errors\n" );
 	}
